requirements.txt
feedparser
requests
python-telegram-bot
python-dotenv
google-generativeai
